Change queries to template strings
