--- packages/venia-concept/src/queries/getCategoryList.graphql
@@ -1,16 +1,0 @@
-# TODO: get only active categories from graphql when it is ready
-query categoryList($id: Int!) {
-    category(id: $id) {
-        children {
-            name
-            url_key
-            url_path
-            image
-            productImagePreview: products(pageSize: 1) {
-                items {
-                    small_image
-                }
-            }
-        }
-    }
-}
