--- packages/venia-concept/src/queries/getCmsBlocks.graphql
@@ -1,8 +1,0 @@
-query cmsBlocks($identifiers: [String]!) {
-    cmsBlocks(identifiers: $identifiers) {
-        items {
-            content
-            identifier
-        }
-    }
-}
