+++ packages/venia-concept/src/queries/getCmsBlocks.js
@@ -1,0 +1,12 @@
+import gql from 'graphql-tag';
+
+export default gql`
+    query cmsBlocks($identifiers: [String]!) {
+        cmsBlocks(identifiers: $identifiers) {
+            items {
+                content
+                identifier
+            }
+        }
+    }
+`;
