--- packages/venia-concept/src/queries/getProductDetail.graphql
@@ -1,56 +1,0 @@
-query productDetail($urlKey: String, $onServer: Boolean!) {
-    productDetail: products(filter: { url_key: { eq: $urlKey } }) {
-        items {
-            sku
-            name
-            price {
-                regularPrice {
-                    amount {
-                        currency
-                        value
-                    }
-                }
-            }
-            description
-            media_gallery_entries {
-                label
-                position
-                disabled
-                file
-            }
-            ... on ConfigurableProduct {
-                configurable_options {
-                    attribute_code
-                    attribute_id
-                    id
-                    label
-                    values {
-                        default_label
-                        label
-                        store_label
-                        use_default_value
-                        value_index
-                    }
-                }
-                variants {
-                    product {
-                        fashion_color
-                        fashion_size
-                        id
-                        media_gallery_entries {
-                            disabled
-                            file
-                            label
-                            position
-                        }
-                        sku
-                        stock_status
-                    }
-                }
-            }
-            meta_title @include(if: $onServer)
-            meta_keyword @include(if: $onServer)
-            meta_description @include(if: $onServer)
-        }
-    }
-}
