+++ packages/venia-concept/src/queries/getProductDetail.js
@@ -1,0 +1,67 @@
+import gql from 'graphql-tag';
+
+/**
+ * As of this writing, there is no single Product query type in the M2.3 schema.
+ * The recommended solution is to use filter criteria on a Products query.
+ * However, the `id` argument is not supported. See
+ * https://github.com/magento/graphql-ce/issues/86
+ * TODO: Replace with a single product query when possible.
+ */
+export default gql`
+    query productDetail($urlKey: String, $onServer: Boolean!) {
+        productDetail: products(filter: { url_key: { eq: $urlKey } }) {
+            items {
+                sku
+                name
+                price {
+                    regularPrice {
+                        amount {
+                            currency
+                            value
+                        }
+                    }
+                }
+                description
+                media_gallery_entries {
+                    label
+                    position
+                    disabled
+                    file
+                }
+                ... on ConfigurableProduct {
+                    configurable_options {
+                        attribute_code
+                        attribute_id
+                        id
+                        label
+                        values {
+                            default_label
+                            label
+                            store_label
+                            use_default_value
+                            value_index
+                        }
+                    }
+                    variants {
+                        product {
+                            fashion_color
+                            fashion_size
+                            id
+                            media_gallery_entries {
+                                disabled
+                                file
+                                label
+                                position
+                            }
+                            sku
+                            stock_status
+                        }
+                    }
+                }
+                meta_title @include(if: $onServer)
+                meta_keyword @include(if: $onServer)
+                meta_description @include(if: $onServer)
+            }
+        }
+    }
+`;
