--- packages/venia-concept/src/queries/urlResolver.graphql
@@ -1,6 +1,0 @@
-query resolveUrl($urlKey: String!) {
-    urlResolver(url: $urlKey) {
-        type
-        id
-    }
-}
