+++ packages/venia-concept/src/queries/urlResolver.js
@@ -1,0 +1,10 @@
+import gql from 'graphql-tag';
+
+export default gql`
+    query resolveUrl($urlKey: String!) {
+        urlResolver(url: $urlKey) {
+            type
+            id
+        }
+    }
+`;
